--- frontend/src/assets/CoursePage/CourseStyle.css
@@ -1,35 +1,0 @@
-  body{
-    background-color: #ffffff;
-    width: 100%;
-  }
-  
-  #root {
-    margin: 0 auto;
-    padding: 2rem;
-    width: 100%;
-    text-align: center;
-  }
-  
-  
-  .card {
-    padding: 2em;
-  }
-  
-  .read-the-docs {
-    color: #888;
-  }
-  
-  .course-list{
-    display: flex;
-    gap: 5px;
-    flex-wrap: wrap;
-  }
-  
-  h1{
-    color: black;
-  }
-  
-  .add-button {
-    background-color: #2196F3;
-    color: #fff;
-  }
