--- frontend/src/assets/componets/CommonComponets/Header/HeaderStyles.css
@@ -1,54 +1,0 @@
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin: 10px 20px 0px 20px;
-    position: relative;
-}
-
-.menu-icon {
-    cursor: pointer;
-    font-size: 1.5rem;
-}
-
-.dropdown-menu {
-    position: absolute;
-    margin-top: 20px;
-    top: 100%;
-    right: 0;
-    background-color: #ffffff;
-    /* border-radius: 8px; */
-    /* box-shadow: 0px 2px 12px rgba(0, 0, 0, 0.2); */
-    /* padding: 10px 0; */
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    transform: translateY(-10px);
-    opacity: 0;
-    transition: opacity 0.3s ease, transform 0.3s ease;
-}
-
-.header .dropdown-menu {
-    transform: translateY(0);
-    opacity: 1;
-}
-
-.dropdown-menu button {
-    background: none;
-    border: none;
-    color: #000000;
-    text-align: left;
-    cursor: pointer;
-    font-size: 1rem;
-    transition: background-color 0.2s ease;
-    padding: 15px;
-}
-
-.dropdown-menu button:hover {
-    background-color: #884545;
-}
-
-.header svg {
-    font-size: 30px;
-    color: #000000;
-}
