+++ frontend/src/assets/componets/ComponentsHome/FrontImg/FrontIng.tsx
@@ -3,7 +3,7 @@ import './FrontImgStyles.css'
 function FrontIng() {
     return (
         <div className='imgDiv'>
-            <img src="../../../../public/img/FrontImg.png" alt="foto" />
+            <img src="./img/FrontImg.png" alt="foto" />
         </div>
     );
 }
