--- frontend/src/assets/componets/ComponentsHome/HomeCards/HomeCard.css
@@ -1,64 +1,0 @@
-.cards{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px;
-}
-
-.card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 210px;
-    height: 110px;
-    border-radius: 25px;
-    padding: 40px;
-    background-color: #ffffff;
-    box-shadow: 0 6px 10px rgba(0, 0, 0, 0.1);
-}
-
-
-.card svg {
-    font-size: 1.5rem;
-    color: #000000;
-    /* margin-top: 5px; */
-    margin-bottom: 10px;
-}
-
-
-.card h5 {
-    font-size: 1.2rem;
-    color: #333;
-    margin: 0;
-}
-
-
-.card p {
-    font-size: 0.9rem;
-    color: #666;
-    margin-top: -2px;
-    margin-bottom: 20px;
-}
-
-
-.card button {
-    padding: 0;
-    font-size: 0.9rem;
-    color: rgb(0, 0, 0);
-    background-color: #ffffff;
-    width: 190px;
-    padding: 5px;
-    border: none;
-    border-radius: 8px;
-    box-shadow: 0 8px 12px rgba(0, 0, 0, 0.1);
-    cursor: pointer;
-    transition: background-color 0.2s;
-    transition: transform 0.2s;
-}
-
-.card button:hover {
-    background-color: #004B84;
-    transform: scale(1.05);
-    color: white;
-}
