--- frontend/src/assets/componets/CourseCard/CourseCardStyle.css
@@ -1,60 +1,0 @@
-/* CourseCard.css */
-.courseCard {
-    width: 300px;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0.2, 0.2);
-    background-color: #fdfdfd;
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-    margin: 10px;
-  }
-  
-  .courseCard:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
-  }
-  
-  .courseCard__title {
-    font-size: 1.5rem;
-    color: #333;
-    margin-bottom: 10px;
-  }
-  
-  .courseCard__description {
-    font-size: 1rem;
-    color: #666;
-    margin-bottom: 15px;
-  }
-  
-  .courseCard__buttons {
-    display: flex;
-    justify-content: space-around;
-  }
-  
-  button {
-    padding: 8px 12px;
-    border-radius: 5px;
-    font-size: 0.9rem;
-    border: none;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-  }
-  
-  .edit-button {
-    background-color: #4CAF50;
-    color: #fff;
-  }
-  
-  .delete-button {
-    background-color: #f44336;
-    color: #fff;
-  }
-  
-  .add-button {
-    background-color: #2196F3;
-    color: #fff;
-  }
-  
-  button:hover {
-    opacity: 0.9;
-  }
