--- frontend/src/assets/componets/CourseFormModal/CourseFormModal.css
@@ -1,30 +1,0 @@
-.modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: #00000099;
-  }
-  
-  .modal-main {
-    position: fixed;
-    background: white;
-    width: 35rem;
-    height: auto;
-    top: 40%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border-radius: 20px;
-  }
-  
-  .modal.display-block {
-    display: block;
-  }
-  
-  .modal.display-none {
-    display: none;
-  }
