--- frontend/src/assets/componets/LoginComponents/LoginForm/LoginPageFormStyles.css
@@ -1,85 +1,0 @@
-.LoginPageContainer {
-    width: 300px;
-    margin: auto;
-    padding: 20px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    text-align: center;
-    font-family: Arial, sans-serif;
-  }
-  
-  .LoginPageContainer h2 {
-    font-size: 20px;
-    margin-bottom: 20px;
-  }
-  
-  .LoginGoogleButton {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-    border: 1px solid #ddd;
-    padding: 8px;
-    border-radius: 4px;
-    cursor: pointer;
-    margin-bottom: 20px;
-  }
-  
-  .LoginGoogleButton img {
-    width: 18px;
-    margin-right: 8px;
-  }
-  
-  .divider {
-    margin: 10px 0;
-    font-size: 12px;
-    color: #666;
-  }
-  
-  form label {
-    display: block;
-    text-align: left;
-    font-size: 14px;
-    margin: 10px 0 5px;
-  }
-  
-  input[type="text"],
-  input[type="password"] {
-    width: 100%;
-    padding: 8px;
-    margin-bottom: 10px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-  }
-  
-  .LoginRememberForgot {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    font-size: 12px;
-  }
-  
-  .ForgotPassword {
-    color: #007bff;
-    text-decoration: none;
-  }
-  
-  .ForgotPassword:hover {
-    text-decoration: underline;
-  }
-  
-  .LoginPageButton {
-    width: 100%;
-    padding: 10px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    font-size: 16px;
-    cursor: pointer;
-  }
-  
-  .LoginPageButton:hover {
-    background-color: #0056b3;
-  }
