+++ frontend/src/assets/page/NoticesPage/NoticesPage.tsx
@@ -1,6 +1,7 @@
 // import React from 'react';
 import './NoticesPageStyles.css'
 import NoticeCard from '../../componets/Cards/NoticeCard/NoticeCard'
+import { FaSearch } from 'react-icons/fa';
 
 
 function NoticesPage() {
